Move override_debian.py to _internal
